Python 3 fixes - Get it installing and running on Python3
Todo:
- Tests still not ported
- So far only tested MySQL collector and hostedgraphite handler

+++ src/collectors/udp/udp.py
@@ -92,7 +92,7 @@ class UDPCollector(diamond.collector.Collector):
                 continue
 
             value = metrics[metric_name]
-            value = self.derivative(metric_name, long(value))
+            value = self.derivative(metric_name, int(value))
 
             # Publish the metric
             self.publish(metric_name, value, 0)
